cleanup of code
