+++ src/xcode/ENA/ENA/Source/Services/PPAC/Model/PPACError.swift
@@ -1,0 +1,12 @@
+////
+// 🦠 Corona-Warn-App
+//
+
+import Foundation
+
+enum PPACError: Error {
+	case generationFailed
+	case deviceNotSupported
+	case timeIncorrect
+	case timeUnverified
+}
