+++ src/xcode/ENA/ENA/Source/Services/PPAC/Model/PPACToken.swift
@@ -1,0 +1,10 @@
+////
+// 🦠 Corona-Warn-App
+//
+
+import Foundation
+
+struct PPACToken {
+	let apiToken: String
+	let deviceToken: String
+}
