+++ src/xcode/ENA/ENA/Source/Services/PPAC/Model/TimestampedToken.swift
@@ -1,0 +1,10 @@
+////
+// 🦠 Corona-Warn-App
+//
+
+import Foundation
+
+struct TimestampedToken: Codable {
+	let token: String
+	let timestamp: Date
+}
