+++ src/xcode/ENA/ENA/Source/Services/PPAC/PPACService.swift
@@ -5,22 +5,6 @@
 import Foundation
 import DeviceCheck
 
-struct TimestampedToken: Codable {
-	let token: String
-	let timestamp: Date
-}
-
-enum PPACError: Error {
-	case generationFailed
-	case deviceNotSupported
-	case timeIncorrect
-	case timeUnverified
-}
-
-struct PPACToken {
-	let apiToken: String
-	let deviceToken: String
-}
 
 protocol PrivacyPreservingAccessControl {
 	func getPPACToken(_ completion: @escaping (Result<PPACToken, PPACError>) -> Void)
